Renamed files for `JoltCustomEmptyShape`
