+++ src/shapes/jolt_custom_empty_shape.hpp
